Relocate `handleDelayedCacheInvalidate`
